% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisPotential.R
\name{ATS}
\alias{ATS}
\title{Compute the ATS Statistic for a Hypothesis Matrix}
\usage{
ATS(X, H)
}
\arguments{
\item{X}{A numeric vector representing the data vector.}

\item{H}{A numeric matrix representing the hypothesis matrix.}
}
\value{
A numeric value representing the ATS statistic.
}
\description{
This function computes the classical ATS (Anova-Type-Statistic) value for a
given hypothesis matrix \code{H} and a data vector \code{X}. The ATS is defined as the
sum of squared values of the product between the hypothesis matrix and the
data vector.
}
\examples{
X <- matrix(stats::runif(10), ncol = 1)
H <- matrix(stats::runif(100), nrow = 10)
ATS(X, H)
}
