% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisPotential.R
\name{WTS}
\alias{WTS}
\title{Compute the Wald-Type-Statistic (WTS) for a Hypothesis Matrix}
\usage{
WTS(X, H, Sigma = NULL)
}
\arguments{
\item{X}{A numeric vector representing the data vector.}

\item{H}{A numeric matrix representing the hypothesis matrix.}

\item{Sigma}{An optional covariance matrix (defaults to an identity matrix).}
}
\value{
A numeric value representing the weighted test statistic.
}
\description{
This function computes the Wald-Type-Statistic (WTS) for a given hypothesis
matrix \code{H} and a data vector\code{X}, optionally using a covariance
matrix \code{Sigma}. For standardization, the Moore-Penrose-inverse of
\code{H\%*\%Sigma\%*\%t(H)} is used.
}
\examples{
X <- matrix(stats::runif(10), ncol = 1)
H <- matrix(stats::runif(100), nrow = 10)
WTS(X, H)
}
