% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisPotential.R
\name{ATS_s}
\alias{ATS_s}
\title{Compute the standardized ATS for a Hypothesis Matrix}
\usage{
ATS_s(X, H, Sigma = NULL)
}
\arguments{
\item{X}{A numeric vector representing the data vector.}

\item{H}{A numeric matrix representing the hypothesis matrix.}

\item{Sigma}{An optional covariance matrix (defaults to an identity matrix).}
}
\value{
A numeric value representing the scaled ATS statistic.
}
\description{
This function computes the standardized ATS (ATS_s) value for a given
hypothesis matrix \code{H} and a data vector \code{X}, optionally using a covariance
matrix \code{Sigma}. The ATS_s is the ATS statistic scaled by the trace#' of the
matrix \code{H\%*\%Sigma\%*\%t(H)}.
}
\examples{
X <- matrix(stats::runif(10), ncol = 1)
H <- matrix(stats::runif(100), nrow = 10)
ATS_s(X, H)
}
