% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centeringCompanion.R
\name{centeringCompanion}
\alias{centeringCompanion}
\title{Generate Companion for the centering Matrix}
\usage{
centeringCompanion(d)
}
\arguments{
\item{d}{Integer. The number of elements which should be compared, which must
be greater than 1.}
}
\value{
A numeric matrix L of size \verb{(d - 1) × d}, which satisfies
\verb{t(L) \\\%*\\\% L = Pd}, where \code{Pd} is the centering matrix \code{diag(d) - 1/d}.
}
\description{
Constructs an \verb{(d-1) × d} matrix that allows to test the usual equality of d
elements, whereto usually the centering matrix \code{Pd} is used.
The companion has a lower number of rows than the centering matrix, but leads
to the same values of the ATS, MATS or WTS.
}
\examples{
L=centeringCompanion(4)
t(L) \%*\% L  # Approximate centering matrix

}
