% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanionHypothesis.R
\name{CompanionHypothesis}
\alias{CompanionHypothesis}
\title{Companion Hypothesis Matrix Transformation}
\usage{
CompanionHypothesis(H, y = NULL, utrapez = 0)
}
\arguments{
\item{H}{A numeric matrix representing the hypothesis. It should be of size
\verb{m x n}, where \code{m} is the number of rows.}

\item{y}{An optional numeric vector, which only has to be specified it is not a zero
vector. It will be transformed alongside the matrix.}

\item{utrapez}{A binary option, which specifies whether the transformed matrix
L should be a upper trapezoidal matrix}
}
\value{
A list with two components:
\itemize{
\item L: The transformed matrix (companion matrix).
\item ytilde: The transformed vector.
}
}
\description{
This function transforms the given hypothesis into a companion hypothesis.
It checks whether the original matrix \code{H} has full row rank. If the matrix
does not have full row rank, the function will proceed to generate a companion matrix
and return the matrix along with a transformed vector \code{ytilde}. If a vector \code{y} is
provided, it will be transformed as well.
}
\details{
If the rank of the matrix \code{H} has full row rank, an error message will be shown,
indicating that the matrix should only be used
if the original matrix has not full row rank.

If the matrix \code{H} has no full row rank, the compact matrix square root of
\code{t(H)\%*\% H} is computed. The vector \code{y} is then transformed
accordingly, and the scale factor is computed to ensure the norm of the
vector remains unchanged after the transformation.
}
\examples{
# Example 1: Matrix with full row rank
H <- matrix(c(1, 2, 3, 4), nrow = 2)
CompanionHypothesis(H)

# Example 2: Matrix with not full row rank and vector y
H <- matrix(c(1, 2, 1, 2), nrow = 2)
y <- c(5, 6)
CompanionHypothesis(H, y)

}
